Consolidate smaller CC files
